refactor(directive): rename query-decorator to util-decorators
